requirements.txt
autopep8==1.4.4
Babel==2.7.0
bleach==3.1.0
bootstrap4==0.1.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.0.13
django-admin-csv==0.2
django-allauth==0.39.1
django-autoslug==1.9.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap4==0.0.8
django-braces==1.13.0
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-croppie==1.3
django-dbbackup==3.2.0
django-email-extras==0.3.4
django-embed-video==1.2.0
django-endless-pagination==2.0
django-extensions==2.1.9
django-filter==2.2.0
django-forms-builder==0.14.0
django-formtools==2.1
django-js-asset==1.2.2
django-model-utils==3.1.2
django-modeladmin-reorder==0.3.1
django-newsletter==0.7
django-nine==0.2.2
django-nonefield==0.3
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
django-smuggler==0.9.1
django-social-share==1.3.2
django-star-ratings==0.7.0
django-taggit==1.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-timed-auth-token==1.3.0
easy-thumbnails==2.6
future==0.15.0
get==2019.4.13
googlemaps==3.0.2
gunicorn==19.9.0
idna==2.8
ldif3==3.1.1
oauthlib==3.0.1
phonenumbers==8.10.14
Pillow==6.2.1
post==2019.4.13
psycopg2==2.8.4
psycopg2-binary==2.8.4
public==2019.4.13
pycodestyle==2.5.0
PyJWT==1.7.1
PyPDF2==1.26.0
pypng==0.0.19
PyQRCode==1.2.1
python-card-me==0.9.3
python-dateutil==2.8.0
python-gnupg==0.4.4
python-resize-image==1.1.19
python3-openid==3.1.0
pytz==2019.1
query-string==2019.4.13
reportlab==3.5.32
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.8.0
six==1.12.0
sorl-thumbnail==12.5.0
sphinx-me==0.3
sqlparse==0.3.0
surlex==0.2.0
swapper==1.1.0
unicodecsv==0.14.1
Unidecode==1.1.0
urllib3==1.25.3
vishap==0.1.5
webencodings==0.5.1
